pyproject.toml
[project]
name = "multimodel-compliance-ingestion-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "azure-identity>=1.25.3",
    "azure-monitor-opentelemetry>=1.8.7",
    "azure-search-documents>=11.6.0",
    "azure-storage-blob>=12.28.0",
    "fastapi>=0.135.3",
    "firecrawl-py>=4.22.0",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.6",
    "langsmith>=0.7.25",
    "opentelemetry-instrumentation-fastapi>=0.61b0",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "redis>=7.4.0",
    "requests>=2.33.1",
    "sqlalchemy>=2.0.49",
    "streamlit>=1.56.0",
    "uvicorn>=0.43.0",
    "yt-dlp>=2026.3.17",
]
